--- a/src/parse/mod.rs
+++ b/src/parse/mod.rs
@@ -4,14 +4,17 @@ pub mod features;
 mod arg_matcher;
 pub mod matches;
 mod parser;
+#[cfg(feature = "validators")]
 mod validator;
 
 pub(crate) use self::{
     arg_matcher::ArgMatcher,
     matches::{MatchedArg, SubCommand, ValueType},
     parser::{Input, ParseResult, Parser},
-    validator::Validator,
 };
+
+#[cfg(feature = "validators")]
+pub(crate) use self::validator::Validator;
 
 pub use self::matches::ArgMatches;
 pub use self::matches::{OsValues, Values};