Validators are not always used, in fact it *might* be fair to say they're rarely used. However they make up a large portion of the binary size. This PR feature gates them with a new feature `validators` which is part of the default feature set.

## Setup

```rust
// src/main.rs
use clap::{App, Arg};
fn main() {
    let matches = App::new("test")
        .arg(Arg::new("foo").short('f').long("foo").default_value("99"))
        .arg(Arg::new("bar").short('b').long("bar"))
        .get_matches();

    let foo = matches.value_of("foo").unwrap();
    println!("foo: {}", foo);
}
```

## Before

**NOTE:**  this is without the `derive` feature

```
cargo bloat --release --crates | head
    Finished release [optimized] target(s) in 0.01s
    Analyzing target/release/fake

 File  .text     Size Crate
 9.3%  58.0% 327.3KiB clap
 4.5%  28.1% 158.4KiB std
 1.7%  10.3%  58.3KiB [Unknown]
 0.1%   0.6%   3.3KiB indexmap
 0.1%   0.5%   3.1KiB strsim
 0.1%   0.3%   1.8KiB termcolor
 0.0%   0.3%   1.6KiB textwrap
 0.0%   0.3%   1.5KiB fake
 0.0%   0.0%      41B os_str_bytes
```
Top 5 largest functions:

```
cargo bloat --release -n 5 | head
    Finished release [optimized] target(s) in 0.01s
    Analyzing target/release/fake

 File  .text     Size     Crate Name
 1.0%   6.2%  35.3KiB      clap clap::parse::parser::Parser::get_matches_with
 0.5%   3.3%  18.4KiB      clap clap::parse::validator::Validator::validate
 0.5%   3.0%  17.1KiB      clap clap::output::help::Help::write_help
 0.5%   3.0%  17.0KiB      clap clap::output::help::Help::write_arg
 0.3%   1.9%  10.7KiB [Unknown] add_ranges
13.0%  80.9% 456.8KiB           And 1216 smaller methods. Use -n N to show more.
16.0% 100.0% 564.6KiB           .text section size, the file size is 3.4MiB
```

## After (when opt-ing out of validators)

```
cargo bloat --release --crates | head
    Finished release [optimized] target(s) in 0.01s
    Analyzing target/release/fake

 File  .text     Size Crate
 7.7%  52.5% 265.4KiB clap
 4.7%  31.9% 161.2KiB std
 1.7%  11.5%  58.3KiB [Unknown]
 0.1%   0.6%   3.3KiB indexmap
 0.1%   0.6%   3.1KiB strsim
 0.1%   0.4%   1.8KiB termcolor
 0.0%   0.3%   1.6KiB textwrap
 0.0%   0.3%   1.5KiB fake
 0.0%   0.0%      41B os_str_bytes
```

Top 5 largest functions:

```
cargo bloat --release -n 5 | head
   Compiling clap v3.0.0-beta.1 (/storage/Projects/clap)
   Compiling fake v0.1.0 (/storage/Projects/fake)
    Finished release [optimized] target(s) in 4.97s
    Analyzing target/release/fake

 File  .text     Size     Crate Name
 1.0%   7.1%  35.8KiB      clap clap::parse::parser::Parser::get_matches_with
 0.5%   3.4%  17.1KiB      clap clap::output::help::Help::write_help
 0.5%   3.4%  17.0KiB      clap clap::output::help::Help::write_arg
 0.3%   2.1%  10.7KiB [Unknown] add_ranges
 0.3%   2.0%  10.3KiB [Unknown] read_line_info
11.8%  80.2% 405.3KiB           And 1157 smaller methods. Use -n N to show more.
14.7% 100.0% 505.0KiB           .text section size, the file size is 3.4MiB
```

Relates to #1365