@kbknapp 2020-05-30
It looks like one of the tests that's failing should be using a different command.

The current command:

```
$ cargo test --target x86_64-unknown-linux-gnu --no-default-features --features std cargo -p clap:3.0.0-beta.1
```

Correct:

```
$ cargo test --target x86_64-unknown-linux-gnu --no-default-features --features "std cargo" -p clap:3.0.0-beta.1
```

Also correct:

```
$ cargo test --target x86_64-unknown-linux-gnu --no-default-features --features std,cargo -p clap:3.0.0-beta.1
```