@kbknapp 2020-05-30
Turns out I'll need to go back to the drawing board. A case of two things named similar, but not doing the same thing. I'll re-open when I've had a chance to *actually* make validators opt-out :wink: