@CreepySkeleton 2020-05-30
`clap::parse::validator::Validator::validate` has very little to do with `.validator()` indeed. In fact, here's _all_ the code that has something to do with `.validator` (not counting derive, tests, debug, and asserts):

https://github.com/clap-rs/clap/blob/229cbbf7a2dc71c2af980164da84b227012cbb6c/src/build/arg/mod.rs#L1912
https://github.com/clap-rs/clap/blob/229cbbf7a2dc71c2af980164da84b227012cbb6c/src/parse/validator.rs#L143-L164

That's it!!! 

I dare to say it doesn't worth a separate feature and it doesn't worth the effort.